# node/api/modules/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Todo API",
        "version": "0.0.1",
        "description": "Use to demonstrate skills",
        "contact": {
            "name": "Todo API",
            "url": "https://gilberthmonge.com/"
        }
    },
    "host": "localhost:8061",
    "basePath": "/api/",
...
    "schemes": [],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "securityDefinitions": {
        "Bearer":{
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
        "Login": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "description": "Email used to create account."
                },
                "password": {
                    "type": "string",
                    "description": "Password used to create account."
                }
            },
            "example": {
                "email": "mail@email.com",
                "password": "TestTodo"
            }
        },
        "Register": {
            "type": "object",
            "properties": {
                "firstname": {
                    "type": "string",
                    "description": "User's first name."
                },
                "lastname": {
                    "type": "string",
                    "description": "User's last name."
                },
                "email": {
                    "type": "string",
                    "description": "Email to create a user account."
                },
                "conf_email": {
                    "type": "string",
                    "description": "To confirm the email."
                },
                "dateOfBirth": {
                    "type": "string",
                    "description": "User's date of birth."
                },
                "password": {
                    "type": "string",
                    "description": "Password used to create account."
                },
                "conf_password": {
                    "type": "string",
                    "description": "To confirm the password."
                }
            },
            "example": {
                "firstname": "Gilberth",
                "lastname": "Monge",
                "email": "gmonge@email.com",
                "conf_email": "gmonge@email.com",
                "dateOfBirth": "1996-02-05",
                "password": "TestTodo",
                "conf_password": "TestTodo"
            }
        },
...
        "SuccessMessage": {
            "type": "object",
            "properties": {
                "state": {
                    "type": "boolean",
                    "description": "Status of request."
                },
                "msg": {
                    "type": "string",
                    "description": "Error if exist."
                },
                "data": {
                    "type": "object",
                    "description": "List the data if they exist."
                }
            },
            "example": {
                "status": 1,
                "msg": "",
                "data": []
            }
        },
        "FailedMessage": {
            "type": "object",
            "properties": {
                "state": {
                    "type": "boolean",
                    "description": "Status of request."
                },
                "msg": {
                    "type": "string",
                    "description": "Error if exist."
                },
                "errors": {
                    "type": "object",
                    "description": "List the errors if they exist."
                }
            },
            "example": {
                "state": false,
                "msg": "Errors found in the request",
                "errors": []
            }
        }
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": ["Authorization"],
                "summary": "Login",
                "description": "Login into API",
                "parameters": [
                    {
                        "in": "body",
                        "name": "data",
                        "required": true,
                        "description": "Data",
                        "schema": {
                            "$ref": "#/definitions/Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return message or data",
                        "schema": {
                            "$ref": "#/definitions/SuccessMessage"
                        }
                    },
                    "401": {
                         "description": "Return error",
                         "schema": {
                             "$ref": "#/definitions/FailedMessage"
                         }
                    },
                    "403": {
                         "description": "Return error",
                         "schema": {
                             "$ref": "#/definitions/FailedMessage"
                         }
                    },
                    "404": {
                        "description": "Return error",
                        "schema": {
                            "$ref": "#/definitions/FailedMessage"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": ["Authorization"],
                "summary": "Register",
                "description": "Register into API",
                "parameters": [
                    {
                        "in": "body",
                        "name": "data",
                        "required": true,
                        "description": "Data",
                        "schema": {
                            "$ref": "#/definitions/Register"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return message or data",
                        "schema": {
                            "$ref": "#/definitions/SuccessMessage"
                        }
                    },
                    "401": {
                         "description": "Return error",
                         "schema": {
                             "$ref": "#/definitions/FailedMessage"
                         }
                    },
                    "403": {
                         "description": "Return error",
                         "schema": {
                             "$ref": "#/definitions/FailedMessage"
                         }
                    },
                    "404": {
                        "description": "Return error",
                        "schema": {
                            "$ref": "#/definitions/FailedMessage"
                        }
                    }
                }
            }
        },
...
    }
}